+++ CHANGELOG.md
@@ -4,6 +4,8 @@ See updating [Changelog example here](https://keepachangelog.com/en/1.0.0/)
 
 ## 0.3.0 (Unreleased)
 
+### Added
+- Added paymentMethod field to Subscription resource
 ### Removed
 
 - Removed a deprecated attribute: persistent_storage_encryption